replace atomic bool with RWMutex to prevent accidental write on closed channel panic
